requirements.txt
annotated-types==0.7.0
anthropic==0.31.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
jiter==0.5.0
MarkupSafe==2.1.5
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
gunicorn==20.1.0